actualizacion de permisos

+++ routes/web.php
@@ -41,28 +41,30 @@ Route::get('order/{id}', 'PedidosController@show'); // Vista de impresion de la 
 Route::get('order-pdf','PedidosController@exportPdf'); // Vista PDF de la orden
 
 
+    Route::get('admin/products', 'ProductController@index'); //listar 
+	Route::get('admin/products/create', 'ProductController@create'); //formulario para crear
+	Route::post('admin/products', 'ProductController@store'); //crear
+	Route::get('admin/products/{id}/edit', 'ProductController@edit'); //form editar
+	Route::post('admin/products/{id}/edit', 'ProductController@update'); //actualizar
+	Route::post('admin/products/{id}/delete', 'ProductController@destroy'); //eliminar
 
-Route::middleware(['auth','admin'])->namespace('Admin')->prefix('admin')->group(function () {
-	Route::get('/products', 'ProductController@index'); //listar 
-	Route::get('/products/create', 'ProductController@create'); //formulario para crear
-	Route::post('/products', 'ProductController@store'); //crear
-	Route::get('/products/{id}/edit', 'ProductController@edit'); //form editar
-	Route::post('/products/{id}/edit', 'ProductController@update'); //actualizar
-	Route::post('/products/{id}/delete', 'ProductController@destroy'); //eliminar
-
-	Route::get('/products/{id}/images', 'ImageController@index'); //listado imagenes 
-	Route::post('/products/{id}/images', 'ImageController@store'); //registrar
-	Route::delete('/products/{id}/images', 'ImageController@destroy'); //eliminar image
-	Route::get('/products/{id}/images/select/{image}', 'ImageController@select'); //destacar 
+	Route::get('admin/products/{id}/images', 'ImageController@index'); //listado imagenes 
+	Route::post('admin/products/{id}/images', 'ImageController@store'); //registrar
+	Route::delete('admin/products/{id}/images', 'ImageController@destroy'); //eliminar image
+	Route::get('admin/products/{id}/images/select/{image}', 'ImageController@select'); //destacar 
 
 	//category
-	Route::get('/categories', 'CategoryController@index'); //listar 
-	Route::get('/categories/create', 'CategoryController@create'); //formulario para crear
-	Route::post('/categories', 'CategoryController@store'); //crear
-	Route::get('/categories/{category}/edit', 'CategoryController@edit'); //form editar
+	Route::get('admin/categories', 'CategoryController@index'); //listar 
+	Route::get('admin/categories/create', 'CategoryController@create'); //formulario para crear
+	Route::post('admin/categories', 'CategoryController@store'); //crear
+	Route::get('admin/categories/{category}/edit', 'CategoryController@edit'); //form editar
 	
-	Route::post('/categories/{category}/edit', 'CategoryController@update'); //actualizar
-	Route::delete('/categories/{category}', 'CategoryController@destroy'); //eliminar
+	Route::post('admin/categories/{category}/edit', 'CategoryController@update'); //actualizar
+	Route::delete('admin/categories/{category}', 'CategoryController@destroy'); //eliminar
 
+
+
+Route::middleware(['auth','admin'])->namespace('Admin')->prefix('admin')->group(function () {
+	
 	
 });
